refactor meetingService, as the file was way too long. Split it in some smaller logical unit
